move android source files into a more sensible package path 'com.hrs.firebase.messaging'
