+++ src/android/com/hrs/firebase/messaging/FCMPluginActivity.java
@@ -1,4 +1,4 @@
-package com.gae.scaffolder.plugin;
+package com.hrs.firebase.messaging;
 
 import android.app.Activity;
 import android.app.NotificationManager;
