+++ src/android/com/hrs/firebase/messaging/interfaces/OnFinishedListener.java
@@ -1,4 +1,4 @@
-package com.gae.scaffolder.plugin.interfaces;
+package com.hrs.firebase.messaging.interfaces;
 
 public interface OnFinishedListener<TResult> {
     void success(TResult result);
