+++ src/android/com/hrs/firebase/messaging/interfaces/TokenListeners.java
@@ -1,4 +1,4 @@
-package com.gae.scaffolder.plugin.interfaces;
+package com.hrs.firebase.messaging.interfaces;
 
 public interface TokenListeners<TSuccess, TError> extends OnFinishedListener<TSuccess> {
     void error(TError message);
